pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Kevin Fortier <kevin.r.fortier@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8"
SQLAlchemy = "^1.4.42"
python-dotenv = "^0.21.0"
mysql-connector-python = "^8.0.31"
mysqlclient = "^2.1.1"
configparser = "^5.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
